bk/part-III/requirement.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
awscli==1.25.88
backcall==0.2.0
bayesian-optimization==1.2.0
betterproto==1.2.5
bidict==0.22.0
black==22.1.0
bleach==4.1.0
botocore==1.27.87
brotlipy==0.7.0
cachetools==5.0.0
certifi==2019.11.28
click==8.0.3
cycler==0.11.0
datasets==2.3.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
dnspython==2.2.1
docutils==0.16
entrypoints==0.4
et-xmlfile==1.1.0
eventlet==0.33.1
executing==0.8.2
filelock==3.6.0
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.5.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.46.0
grpclib==0.4.2
h2==4.1.0
hpack==4.0.0
httplib2==0.20.4
huggingface-hub==0.4.0
hyperframe==6.0.1
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.1
jmespath==1.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kafka-python==2.0.2
kaleido==0.2.1
kiwisolver==1.3.2
Markdown==3.3.7
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mizani==0.7.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
msgpack==1.0.3
multidict==6.0.2
multiprocess==0.70.13
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
palettable==3.3.0
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
patchworklib==0.4.2
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
plotly==5.9.0
plotnine==0.9.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.11.2
pyrsistent==0.18.1
python-engineio==4.3.2
python-socketio==5.6.0
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
rdflib==6.1.1
rdkit-pypi==2022.3.1
regex==2022.3.15
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.7.2
s3transfer==0.6.0
sacremoses==0.0.49
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
stack-data==0.1.4
statsmodels==0.13.2
stringcase==1.2.0
tenacity==8.0.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-metadata==1.7.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
tokenizers==0.10.3
tomli==2.0.0
torch==1.10.2
torchaudio==0.10.2
torchmetrics==0.7.3
torchvision==0.11.3
tornado==6.1
traitlets==5.1.1
transformers==4.12.5
typing_extensions==4.0.1
uritemplate==4.1.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.2
widgetsnbextension==3.5.2
xgboost==1.6.2
xxhash==3.0.0
yarl==1.7.2
zipp==3.7.0
